fix: bugs with basic screen
