+++ force-app/main/default/lwc/basicScreen/basicScreenHelper.js
@@ -1,0 +1,11 @@
+import { surveyFields } from "c/fieldService";
+
+const createSurveyDisplayedMap = (objectList) => {
+    return objectList.map((element) => {
+        return { label: element[surveyFields.NAME], value: element[surveyFields.ID] };
+    });
+}
+
+export {
+    createSurveyDisplayedMap
+}
